Final Project - IBM Data Science Certification
In this project I'll analyze the principal venues in each neighbourhood in CABA (Ciudad Autónoma de Buenos Aires),the capital and largest city of Argentina, my country.
The propouse of this analysis is to cluster the neighbourhoods in order to choose the best one to start a new business. For example, a new restaurant.
For it, I'll use a list of the Neighbourhoods names, with geopy I'll get the coordinates for each one, and then with the Foursquare API I'll get the principal venues in each neighbour.
Once I've all that information, I'll try a K-means model to cluster the neighbourhoods.